[COST-3685] Move post processor logic into its own provider specific class.  (#4290)
* Restructure to create post processor class.
---------
