+++ koku/masu/test/util/oci/test_common.py
@@ -7,7 +7,6 @@ import shutil
 import tempfile
 from datetime import datetime
 
-import pandas as pd
 from dateutil.relativedelta import relativedelta
 from faker import Faker
 from tenant_schemas.utils import schema_context
@@ -19,7 +18,7 @@ from masu.external.date_accessor import DateAccessor
 from masu.test import MasuTestCase
 from masu.util.oci import common as utils
 from reporting.models import OCICostEntryBill
-from reporting.provider.oci.models import TRINO_REQUIRED_COLUMNS
+
 
 # the cn endpoints aren't supported by moto, so filter them out
 OCI_REGIONS = list(filter(lambda reg: not reg.startswith("cn-"), OCI_REGIONS))
@@ -143,253 +142,6 @@ class TestOCIUtils(MasuTestCase):
             bill_ids = [str(bill.id) for bill in bills]
 
         self.assertEqual(bill_ids, expected_bill_ids)
-
-    def test_oci_post_processor(self):
-        """Test that missing columns in a report end up in the data frame."""
-        column_one = "column_one"
-        column_two = "column_two"
-        column_three = "column-three"
-        column_four = "resourceTags/User:key"
-        data = {column_one: [1, 2], column_two: [3, 4], column_three: [5, 6], column_four: ["value_1", "value_2"]}
-        data_frame = pd.DataFrame.from_dict(data)
-
-        processed_data_frame = utils.oci_post_processor(data_frame)
-        if isinstance(processed_data_frame, tuple):
-            processed_data_frame, df_tag_keys, _ = processed_data_frame
-            self.assertIsInstance(df_tag_keys, set)
-
-        columns = list(processed_data_frame)
-
-        self.assertIn(column_one, columns)
-        self.assertIn(column_two, columns)
-        self.assertIn(column_three.replace("-", "_"), columns)
-        self.assertNotIn(column_four, columns)
-        self.assertIn("tags", columns)
-        for column in TRINO_REQUIRED_COLUMNS:
-            self.assertIn(column.replace("-", "_").replace("/", "_").replace(":", "_").lower(), columns)
-
-    def test_oci_generate_daily_data_usage(self):
-        """Test that we aggregate data at a daily level."""
-        usageamount = random.randint(1, 10)
-        data = [
-            {
-                "lineitem_referenceno": "id1",
-                "lineitem_tenantid": "my-tenant",
-                "lineitem_intervalusagestart": datetime(2021, 6, 7, 11, 24, 0),
-                "lineitem_intervalusageend": datetime(2021, 6, 7, 11, 24, 0),
-                "product_service": "service",
-                "product_resource": "resource",
-                "product_compartmentid": "compart_id",
-                "product_compartmentname": "compart_name",
-                "product_region": "my-region",
-                "product_availabilitydomain": "my-domain",
-                "product_resourceid": "id4",
-                "usage_consumedquantity": usageamount,
-                "usage_billedquantity": "3",
-                "usage_billedquantityoverage": "0",
-                "usage_consumedquantityunits": "BYTES",
-                "usage_consumedquantitymeasure": "BYTES",
-                "cost_subscriptionid": "subscript_id",
-                "cost_productsku": "sku",
-                "product_description": "storage",
-                "lineitem_iscorrection": "",
-                "lineitem_backreferenceno": "",
-                "tags": "",
-            },
-            {
-                "lineitem_referenceno": "id1",
-                "lineitem_tenantid": "my-tenant",
-                "lineitem_intervalusagestart": datetime(2021, 6, 7, 12, 24, 0),  # different hour, same day
-                "lineitem_intervalusageend": datetime(2021, 6, 7, 12, 24, 0),  # different hour, same day
-                "product_service": "service",
-                "product_resource": "resource",
-                "product_compartmentid": "compart_id",
-                "product_compartmentname": "compart_name",
-                "product_region": "my-region",
-                "product_availabilitydomain": "my-domain",
-                "product_resourceid": "id4",
-                "usage_consumedquantity": usageamount,
-                "usage_billedquantity": "3",
-                "usage_billedquantityoverage": "0",
-                "usage_consumedquantityunits": "BYTES",
-                "usage_consumedquantitymeasure": "BYTES",
-                "cost_subscriptiond": "subscript_id",
-                "cost_productsku": "sku",
-                "product_description": "storage",
-                "lineitem_iscorrection": "",
-                "lineitem_backreferenceno": "",
-                "tags": "",
-            },
-            {
-                "lineitem_referenceno": "id1",
-                "lineitem_tenantid": "my-tenant",
-                "lineitem_intervalusagestart": datetime(2021, 6, 8, 12, 24, 0),  # different hour, same day
-                "lineitem_intervalusageend": datetime(2021, 6, 8, 12, 24, 0),  # different hour, same day
-                "product_service": "service",
-                "product_resource": "resource",
-                "product_compartmentid": "compart_id",
-                "product_compartmentname": "compart_name",
-                "product_region": "my-region",
-                "product_availabilitydomain": "my-domain",
-                "product_resourceid": "id4",
-                "usage_consumedquantity": usageamount,
-                "usage_billedquantity": "3",
-                "usage_billedquantityoverage": "0",
-                "usage_consumedquantityunits": "BYTES",
-                "usage_consumedquantitymeasure": "BYTES",
-                "cost_subscriptionid": "subscript_id",
-                "cost_productsku": "sku",
-                "product_description": "storage",
-                "lineitem_iscorrection": "",
-                "lineitem_backreferenceno": "",
-                "tags": "",
-            },
-        ]
-
-        df = pd.DataFrame(data)
-
-        daily_df = utils.oci_generate_daily_data(df)
-
-        first_day = daily_df[daily_df["lineitem_intervalusagestart"] == "2021-06-07"]
-        second_day = daily_df[daily_df["lineitem_intervalusagestart"] == "2021-06-08"]
-
-        # Assert that there is only 1 record per day
-        self.assertEqual(first_day.shape[0], 1)
-        self.assertEqual(second_day.shape[0], 1)
-
-        self.assertTrue((first_day["usage_consumedquantity"] == usageamount * 2).bool())
-
-        self.assertTrue((second_day["usage_consumedquantity"] == usageamount).bool())
-
-    def test_oci_generate_daily_data_cost(self):
-        """Test that we aggregate data at a daily level."""
-        cost = random.randint(1, 10)
-        data = [
-            {
-                "lineitem_referenceno": "id1",
-                "lineitem_tenantid": "my-tenant",
-                "lineitem_intervalusagestart": datetime(2021, 6, 7, 11, 24, 0),
-                "lineitem_intervalusageend": datetime(2021, 6, 7, 11, 24, 0),
-                "product_service": "service",
-                "product_resource": "resource",
-                "product_compartmentid": "compart_id",
-                "product_compartmentname": "compart_name",
-                "product_region": "my-region",
-                "product_availabilitydomain": "my-domain",
-                "product_resourceid": "id4",
-                "usage_billedquantity": "3",
-                "usage_billedquantityoverage": "0",
-                "usage_consumedquantityunits": "BYTES",
-                "usage_consumedquantitymeasure": "BYTES",
-                "cost_subscriptionid": "subscript_id",
-                "cost_productsku": "sku",
-                "product_description": "storage",
-                "cost_unitprice": "10",
-                "cost_unitpriceoverage": "10",
-                "cost_mycost": cost,
-                "cost_mycostoverage": "0",
-                "cost_currencycode": "USD",
-                "cost_billingunitreadable": "",
-                "cost_skuunitdescription": "unitdescription",
-                "cost_overageflag": "",
-                "lineitem_iscorrection": "",
-                "lineitem_backreferenceno": "",
-                "tags": "",
-            },
-            {
-                "lineitem_referenceno": "id1",
-                "lineitem_tenantid": "my-tenant",
-                "lineitem_intervalusagestart": datetime(2021, 6, 7, 12, 24, 0),  # different hour, same day
-                "lineitem_intervalusageend": datetime(2021, 6, 7, 12, 24, 0),  # different hour, same day
-                "product_service": "service",
-                "product_resource": "resource",
-                "product_compartmentid": "compart_id",
-                "product_compartmentname": "compart_name",
-                "product_region": "my-region",
-                "product_availabilitydomain": "my-domain",
-                "product_resourceid": "id4",
-                "usage_billedquantity": "3",
-                "usage_billedquantityoverage": "0",
-                "usage_consumedquantityunits": "BYTES",
-                "usage_consumedquantitymeasure": "BYTES",
-                "cost_subscriptiond": "subscript_id",
-                "cost_productsku": "sku",
-                "product_description": "storage",
-                "cost_unitprice": "10",
-                "cost_unitpriceoverage": "10",
-                "cost_mycost": cost,
-                "cost_mycostoverage": "0",
-                "cost_currencycode": "USD",
-                "cost_billingunitreadable": "",
-                "cost_skuunitdescription": "unitdescription",
-                "cost_overageflag": "",
-                "lineitem_iscorrection": "",
-                "lineitem_backreferenceno": "",
-                "tags": "",
-            },
-            {
-                "lineitem_referenceno": "id1",
-                "lineitem_tenantid": "my-tenant",
-                "lineitem_intervalusagestart": datetime(2021, 6, 8, 12, 24, 0),  # different hour, same day
-                "lineitem_intervalusageend": datetime(2021, 6, 8, 12, 24, 0),  # different hour, same day
-                "product_service": "service",
-                "product_resource": "resource",
-                "product_compartmentid": "compart_id",
-                "product_compartmentname": "compart_name",
-                "product_region": "my-region",
-                "product_availabilitydomain": "my-domain",
-                "product_resourceid": "id4",
-                "usage_billedquantity": "3",
-                "usage_billedquantityoverage": "0",
-                "usage_consumedquantityunits": "BYTES",
-                "usage_consumedquantitymeasure": "BYTES",
-                "cost_subscriptionid": "subscript_id",
-                "cost_productsku": "sku",
-                "product_description": "storage",
-                "cost_unitprice": "10",
-                "cost_unitpriceoverage": "10",
-                "cost_mycost": cost,
-                "cost_mycostoverage": "0",
-                "cost_currencycode": "USD",
-                "cost_billingunitreadable": "",
-                "cost_skuunitdescription": "unitdescription",
-                "cost_overageflag": "",
-                "lineitem_iscorrection": "",
-                "lineitem_backreferenceno": "",
-                "tags": "",
-            },
-        ]
-
-        df = pd.DataFrame(data)
-
-        daily_df = utils.oci_generate_daily_data(df)
-
-        first_day = daily_df[daily_df["lineitem_intervalusagestart"] == "2021-06-07"]
-        second_day = daily_df[daily_df["lineitem_intervalusagestart"] == "2021-06-08"]
-
-        # Assert that there is only 1 record per day
-        self.assertEqual(first_day.shape[0], 1)
-        self.assertEqual(second_day.shape[0], 1)
-
-        self.assertTrue((first_day["cost_mycost"] == cost * 2).bool())
-
-        self.assertTrue((second_day["cost_mycost"] == cost).bool())
-
-    def test_oci_post_processor_empty_tags(self):
-        """Test that missing columns in a report end up in the data frame."""
-        column_one = "column_one"
-        column_two = "column_two"
-        column_three = "column-three"
-        column_four = "resourceTags/System:key"
-        data = {column_one: [1, 2], column_two: [3, 4], column_three: [5, 6], column_four: ["value_1", "value_2"]}
-        data_frame = pd.DataFrame.from_dict(data)
-
-        processed_data_frame = utils.oci_post_processor(data_frame)
-        if isinstance(processed_data_frame, tuple):
-            processed_data_frame, df_tag_keys, _ = processed_data_frame
-            self.assertIsInstance(df_tag_keys, set)
-
-        self.assertFalse(processed_data_frame["tags"].isna().values.any())
 
     def test_detect_type(self):
         "Test that we detect the correct report type from csv"
