+++ koku/masu/util/common.py
@@ -31,166 +31,6 @@ from masu.external import POLL_INGEST
 
 LOG = logging.getLogger(__name__)
 
-CSV_ALT_COLUMNS = {
-    "AWS": (
-        "bill_billing_entity",
-        "bill_bill_type",
-        "bill_payer_account_id",
-        "bill_billing_period_start_date",
-        "bill_billing_period_end_date",
-        "bill_invoice_id",
-        "line_item_line_item_type",
-        "line_item_usage_account_id",
-        "line_item_usage_start_date",
-        "line_item_usage_end_date",
-        "line_item_product_code",
-        "line_item_usage_type",
-        "line_item_operation",
-        "line_item_availability_zone",
-        "line_item_resource_id",
-        "line_item_usage_amount",
-        "line_item_normalization_factor",
-        "line_item_normalized_usage_amount",
-        "line_item_currency_code",
-        "line_item_unblended_rate",
-        "line_item_unblended_cost",
-        "line_item_blended_rate",
-        "line_item_blended_cost",
-        "savings_plan_savings_plan_effective_cost",
-        "line_item_tax_type",
-        "pricing_public_on_demand_cost",
-        "pricing_public_on_demand_rate",
-        "reservation_amortized_upfront_fee_for_billing_period",
-        "reservation_amortized_upfront_cost_for_usage",
-        "reservation_recurring_fee_for_usage",
-        "reservation_unused_quantity",
-        "reservation_unused_recurring_fee",
-        "pricing_term",
-        "pricing_unit",
-        "product_sku",
-        "product_product_name",
-        "product_product_family",
-        "product_servicecode",
-        "product_region",
-        "product_instance_type",
-        "product_memory",
-        "product_vcpu",
-        "reservation_number_of_reservations",
-        "reservation_units_per_reservation",
-        "reservation_start_time",
-        "reservation_end_time",
-    ),
-}
-
-CSV_REQUIRED_COLUMNS = {
-    "AWS": (
-        "bill/BillingEntity",
-        "bill/BillType",
-        "bill/PayerAccountId",
-        "bill/BillingPeriodStartDate",
-        "bill/BillingPeriodEndDate",
-        "bill/InvoiceId",
-        "lineItem/LineItemType",
-        "lineItem/UsageAccountId",
-        "lineItem/UsageStartDate",
-        "lineItem/UsageEndDate",
-        "lineItem/ProductCode",
-        "lineItem/UsageType",
-        "lineItem/Operation",
-        "lineItem/AvailabilityZone",
-        "lineItem/ResourceId",
-        "lineItem/UsageAmount",
-        "lineItem/NormalizationFactor",
-        "lineItem/NormalizedUsageAmount",
-        "lineItem/CurrencyCode",
-        "lineItem/UnblendedRate",
-        "lineItem/UnblendedCost",
-        "lineItem/BlendedRate",
-        "lineItem/BlendedCost",
-        "savingsPlan/SavingsPlanEffectiveCost",
-        "lineItem/TaxType",
-        "pricing/publicOnDemandCost",
-        "pricing/publicOnDemandRate",
-        "reservation/AmortizedUpfrontFeeForBillingPeriod",
-        "reservation/AmortizedUpfrontCostForUsage",
-        "reservation/RecurringFeeForUsage",
-        "reservation/UnusedQuantity",
-        "reservation/UnusedRecurringFee",
-        "pricing/term",
-        "pricing/unit",
-        "product/sku",
-        "product/ProductName",
-        "product/productFamily",
-        "product/servicecode",
-        "product/region",
-        "product/instanceType",
-        "product/memory",
-        "product/vcpu",
-        "reservation/ReservationARN",
-        "reservation/NumberOfReservations",
-        "reservation/UnitsPerReservation",
-        "reservation/StartTime",
-        "reservation/EndTime",
-    ),
-    "Azure": (
-        "SubscriptionGuid",
-        "ResourceGroup",
-        "ResourceLocation",
-        "UsageDateTime",
-        "MeterCategory",
-        "MeterSubcategory",
-        "MeterId",
-        "MeterName",
-        "MeterRegion",
-        "UsageQuantity",
-        "ResourceRate",
-        "PreTaxCost",
-        "ConsumedService",
-        "ResourceType",
-        "InstanceId",
-        "Tags",
-        "OfferId",
-        "AdditionalInfo",
-        "ServiceInfo1",
-        "ServiceInfo2",
-        "ServiceName",
-        "ServiceTier",
-        "Currency",
-        "UnitOfMeasure",
-    ),
-    "GCP": (
-        "billing_account_id",
-        "service.id",
-        "service.description",
-        "sku.id",
-        "sku.description",
-        "usage_start_time",
-        "usage_end_time",
-        "project.id",
-        "project.name",
-        "project.labels",
-        "project.ancestry_numbers",
-        "labels",
-        "system_labels",
-        "location.location",
-        "location.country",
-        "location.region",
-        "location.zone",
-        "export_time",
-        "cost",
-        "currency",
-        "currency_conversion_rate",
-        "usage.amount",
-        "usage.unit",
-        "usage.amount_in_pricing_units",
-        "usage.pricing_unit",
-        "credits",
-        "invoice.month",
-        "cost_type",
-        "partition_date",
-    ),
-}
-
 
 def extract_uuids_from_string(source_string):
     """
@@ -485,24 +325,29 @@ def batch(iterable, start=0, stop=None, _slice=1):
 
 
 def create_enabled_keys(schema, enabled_keys_model, enabled_keys):
-    LOG.info("Creating enabled tag key records")
+    """
+    Creates enabled key records.
+    """
+
+    if not enabled_keys:
+        LOG.info("No enabled keys found")
+        return
+    LOG.info(f"Creating enabled key records: {str(enabled_keys_model._meta.model_name)}.")
     changed = False
 
-    if enabled_keys:
-        with schema_context(schema):
-            new_keys = list(set(enabled_keys) - {k.key for k in enabled_keys_model.objects.all()})
-            if new_keys:
-                changed = True
-                # Processing in batches for increased efficiency
-                for batch_num, new_batch in enumerate(batch(new_keys, _slice=500)):
-                    batch_size = len(new_batch)
-                    LOG.info(f"Create batch {batch_num + 1}: batch_size {batch_size}")
-                    for ix in range(batch_size):
-                        new_batch[ix] = enabled_keys_model(key=new_batch[ix])
-                    enabled_keys_model.objects.bulk_create(new_batch, ignore_conflicts=True)
-
+    with schema_context(schema):
+        new_keys = list(set(enabled_keys) - {k for k in enabled_keys_model.objects.values_list("key", flat=True)})
+        if new_keys:
+            changed = True
+            # Processing in batches for increased efficiency
+            for batch_num, new_batch in enumerate(batch(new_keys, _slice=500)):
+                batch_size = len(new_batch)
+                LOG.info(f"Create batch {batch_num + 1}: batch_size {batch_size}")
+                for ix in range(batch_size):
+                    new_batch[ix] = enabled_keys_model(key=new_batch[ix])
+                enabled_keys_model.objects.bulk_create(new_batch, ignore_conflicts=True)
     if not changed:
-        LOG.info("No enabled keys added.")
+        LOG.info("No enabled keys added")
 
     return changed
